EIN-4500-implementere-innsynskrav-i-nytt-api

+++ src/main/java/no/einnsyn/apiv3/entities/dokumentbeskrivelse/DokumentbeskrivelseRepository.java
@@ -4,5 +4,5 @@ import no.einnsyn.apiv3.entities.EinnsynRepository;
 import no.einnsyn.apiv3.entities.dokumentbeskrivelse.models.Dokumentbeskrivelse;
 
 public interface DokumentbeskrivelseRepository
-        extends EinnsynRepository<Dokumentbeskrivelse, Long> {
+    extends EinnsynRepository<Dokumentbeskrivelse, Integer> {
 }
